Homapage fixed and component, styles setup completed..
