+++ Frontend/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 // import LoginPage from './LoginPage';
-import WelcomePage from './WelcomePage';
-import HomePage from './HomePage';
+import WelcomePage from './Components/WelcomePage';
+import HomePage from './Components/HomePage';
 
 function App() {
   return (
