--- Frontend/src/Navbar.css
@@ -1,79 +1,0 @@
-/* Navbar.css */
-/* .navbar {
-  background-color: var(--white);
-  padding: 1rem 2rem;
-  position: sticky;
-  top: 0;
-  z-index: 999;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-}
-
-.navbar .container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.navbar-brand img {
-  height: 40px;
-}
-
-.hamburger {
-  background: none;
-  border: none;
-  cursor: pointer;
-  display: none;
-}
-
-.navbar-collapse {
-  display: flex;
-}
-
-.navbar-collapse.open {
-  display: block;
-}
-
-.navbar-nav {
-  display: flex;
-  gap: 1.5rem;
-  list-style: none;
-  padding: 0;
-} */
-
-/* .nav-item .nav-link {
-  text-decoration: none;
-  color: var(--black);
-  font-weight: 600;
-  transition: color 0.3s ease;
-} */ */
-
-.navbar-collapse {
-  display: none;
-  flex-direction: column;
-  position: absolute;
-  top: 70px;
-  right: 20px;
-  background-color: var(--white);
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.2);
-  padding: 1rem;
-}
-
-/* .nav-item .nav-link:hover {
-  color: var(--light-blue);
-  background-color: transparent;
-} */
-
-/* @media (max-width: 768px) {
-  .hamburger {
-    display: block;
-  }
-
-  .navbar-collapse.open {
-    display: flex;
-  }
-
-  .navbar-nav {
-    flex-direction: column;
-    gap: 1rem;
-  }
-}
