+++ Frontend/src/Styles/HomePage.css
@@ -52,10 +52,10 @@ img {
 .container {
   width: 100%;
   margin: 0 auto;
-  padding: 1rem;
+  padding: 5rem;
   background-color: white;
   box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
-  border-radius: 10px;
+  border-radius: none;
 }
 
 .bg-blue {
@@ -102,13 +102,14 @@ h3 {
 }
 
 .py {
-  padding: 0.9rem 0;
+  padding-bottom: 1rem;
 }
 
 .grid {
   display: grid;
   grid-template-columns: repeat(2, 1fr);
   gap: 2rem;
+  padding: 3rem;
   align-items: center;
 }
 
@@ -182,18 +183,30 @@ h3 {
 .navbar {
   display: flex;
   background-color: var(--white);
+  justify-content: space-between;
   padding: 1rem 2rem;
-  position: sticky;
+  position: fixed;
   top: 0;
   z-index: 1000;
   box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
 }
 
+/* .navbar-collapse {
+  display: none;
+  flex-direction: column;
+  position: absolute;
+  top: 70px;
+  right: 20px;
+  background-color: var(--white);
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.2);
+  padding: 1rem;
+} */
+/* 
 .navbar .container {
   display: flex;
   justify-content: space-between;
   align-items: center;
-}
+} */
 
 .navbar-brand img {
   height: 40px;
@@ -328,16 +341,24 @@ h3 {
   overflow: hidden;
   position: relative;
   background-color: var(--light-grey);
-  -webkit-box-align: end;
-  -ms-flex-align: end;
-  align-items: flex-end;
-}
-
-.doc-panel-item .img {
   max-width: 357px;
   margin-left: auto;
   margin-right: auto;
+  border-radius: 10px;
 }
+
+.doc-panel-item .img img {
+  width: initial;
+  height: 100%;
+  object-fit: inherit;
+  transition: transform 0.5s ease, box-shadow 0.5s ease;
+}
+
+.doc-panel-item .img:hover img {
+  transform: scale(1.1);
+  box-shadow: 0 8px 20px rgba(0, 0, 0, 0.3);
+}
+
 .doc-panel-item .info {
   position: absolute;
   bottom: 0;
